added control class
